11.4下午

+++ src/main/java/com/zyyglxt/service/impl/HospServiceImpl.java
@@ -3,7 +3,11 @@ package com.zyyglxt.service.impl;
 import com.zyyglxt.dao.HospDOMapper;
 import com.zyyglxt.dataobject.HospDO;
 import com.zyyglxt.dataobject.HospDOKey;
+import com.zyyglxt.error.BusinessException;
+import com.zyyglxt.error.EmBusinessError;
 import com.zyyglxt.service.IHospService;
+import com.zyyglxt.validator.ValidatorImpl;
+import com.zyyglxt.validator.ValidatorResult;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,20 +24,33 @@ import java.util.UUID;
 public class HospServiceImpl implements IHospService {
     @Resource
     HospDOMapper hospDOMapper;
+    @Autowired
+    private ValidatorImpl validator;
 
     @Override
     public int addHosp(HospDO hospDO) {
-        hospDO.setItemcode(UUID.randomUUID().toString());
+        ValidatorResult result = validator.validate(hospDO);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.insertSelective(hospDO);
     }
 
     @Override
     public int updateHosp(HospDO hospDO) {
+        ValidatorResult result = validator.validate(hospDO);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.updateByPrimaryKeySelective(hospDO);
     }
 
     @Override
     public int deleteHosp(HospDOKey hospDOKey) {
+        ValidatorResult result = validator.validate(hospDOKey);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.deleteByPrimaryKey(hospDOKey);
     }
 
@@ -47,6 +64,9 @@ public class HospServiceImpl implements IHospService {
      */
     @Override
     public List<HospDO> searchHosp(String keyWord) {
+        if(keyWord.isEmpty()){
+            throw new BusinessException("关键字不能为空", EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return hospDOMapper.searchHosp(keyWord);
     }
